requirements.txt
autopep8==1.5.6
pycodestyle==2.7.0
python-decouple==3.4
toml==0.10.2
